adding support for torch compile
