+++ tests/kernels/moe/test_new_triton_moe.py
@@ -22,7 +22,7 @@ def forward_cuda_ref(x, w1, w2,
                     num_expert_group: Optional[int] = None,
                     global_num_experts: int = -1,
                     expert_map: Optional[torch.Tensor] = None,
-                    custom_routing_function: Optional[Callable] = None,
+                    # custom_routing_function: Optional[Callable] = None,   // custom routing kernel is not supported due to pytorch ops constraint
                     scoring_func: str = "softmax",
                     e_score_correction_bias: Optional[torch.Tensor] = None,
                     apply_router_weight_on_input: bool = False,
@@ -141,7 +141,7 @@ def forward_cuda_exp(x, w1, w2,
                     num_expert_group: Optional[int] = None,
                     global_num_experts: int = -1,
                     expert_map: Optional[torch.Tensor] = None,
-                    custom_routing_function: Optional[Callable] = None,
+                    # custom_routing_function: Optional[Callable] = None,
                     scoring_func: str = "softmax",
                     e_score_correction_bias: Optional[torch.Tensor] = None,
                     apply_router_weight_on_input: bool = False,
