+++ vllm/model_executor/layers/fused_moe/__init__.py
@@ -41,11 +41,14 @@ if HAS_TRITON:
     from vllm.model_executor.layers.fused_moe.fused_moe import (
         TritonExperts, fused_experts, fused_moe, fused_topk,
         get_config_file_name, grouped_topk)
+    from vllm.model_executor.layers.fused_moe.triton_kernels_moe import (
+        forward_cuda_triton)
 
     __all__ += [
         "fused_moe",
         "fused_topk",
         "fused_experts",
+        "forward_cuda_triton"
         "get_config_file_name",
         "grouped_topk",
         "cutlass_moe_fp8",
